separate Weather Enum into Swift File
