+++ AddressBookApp/AddressBookApp/TableViewController.swift
@@ -66,28 +66,3 @@ class TableViewController: UITableViewController {
         weatherDescription.font = UIFont.boldSystemFont(ofSize: 20)
     }
 }
-
-enum Weather: Int, CaseIterable {
-    case cloudy = 0
-    case rainny = 1
-    case snowy = 2
-    case sunny = 3
-    
-    func getText() -> String {
-        switch self {
-        case .cloudy: return "오늘 날씨를 구름이낍니다."
-        case .rainny: return "오늘 날씨는 흐림입니다."
-        case .snowy: return "오늘 날씨는 눈이옵니다."
-        case .sunny: return "오늘 날씨는 화창합니다."
-        }
-    }
-    
-    func getImageName() -> String {
-        switch self {
-        case .cloudy: return "weather-cloudy"
-        case .rainny: return "weather-rainny"
-        case .snowy: return "weather-snowy"
-        case .sunny: return "weather-sunny"
-        }
-    }
-}
